Refactor: home_page divided into function & UI
